Added a BetterDefaultQueryFilter to not fail when a polygon has no flag set at all (which is the default for the NavMeshDataCreateParamsBuilder), because I feel "regular ground" doesn't deserve a flag. The downside is that you can't exclude this "regular ground" with the excludeFlags though and hence would need the DefaultQueryFilter which automatically discards those polygons for you.
So we had the decision to make between adding such a filter and recommending that to everyone or setting the flag to 0x1 for every polygon as default. That decision is still up to you, you can use the withFlags() methods on said Builder and use the DefaultQueryFilter to have the recast4j intended behavior, but personally I feel more satisfied with regular ground being nothing but regular ground, I don't need special information or filtering in most cases anyway.